Replace SQLiteWriter d_columns with metadata call
The d_columns field is removed. Instead, to find out whether some column
or table exists, the sqlite3_table_column_metadata function is called.

This resolves an issue where multiple connections are open to the SQLite
database and calls to addValue cause columns to be added to a table that
is also used by others. Before, these new columns would only be added to
the d_columns list of the connection that created them, meaning that all
others would be out of sync, and stay that way. If they added a row that
had the new column, they would attempt an ALTER TABLE, which would fail.

This change also means that we automatically follow SQLite when it comes
to case sensitivity, so this should also fix #9.

Finally, the code is slightly simpler.

The getSchema method on MiniSQLite is now unused by the code itself, but
since the method is public it is kept in case it is used by API users.